Add NFCSTATUS_FAILED
On some handsets, there is a configuration of hardware that causes a
conflict with HCE.

Primarily it is something in the SWP configuration with a SWP sim card.
It appears to be a PN544 register setting that blocks HCE type A only
from initialization.
This patch checks for this conflict and disables HCE for those
configurations.
The better way would be to understand how to set the PN544 registers to
unblock the conflict, but this is a sufficient work around for this
minority of configurations.

Change-Id: I64d79acaa7eb58e15f1fb4bbe298883036f6d45a